# Experiment: Load theme's block patterns concurrently using stream_select()