@dmsnell 2023-10-13
Closing because the immediate need is resolved. It might be worth exploring further the concurrent I/O helper in places where I/O lag is a measured and substantive obstacle.